Extract Theme Toggle Button to a React component
